% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_series.R
\name{available_series}
\alias{available_series}
\title{Available Series}
\usage{
available_series(language = c("en", "br"), label = TRUE)
}
\arguments{
\item{language}{A character string specifying the language. Available
options are English (`"en"`, default) and Brazilian Portuguese (`"br"`).}

\item{label}{Logical. If `TRUE` (default), variable labels are added to
the output data frame using `sjlabelled::set_label()`.}
}
\value{
A data frame containing the Ipeadata code, name, theme, source,
  frequency, last update, and activity status of available series.
}
\description{
Returns a list of available series from the Ipeadata API.
}
\note{
The original language of the available series names is preserved.
}
