% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipeadata.R
\name{ipeadata}
\alias{ipeadata}
\title{Data for Requested Series}
\usage{
ipeadata(code, language = c("en", "br"), label = TRUE, quiet = FALSE)
}
\arguments{
\item{code}{A character vector of Ipeadata series codes.}

\item{language}{A character string specifying the language. Available
options are English (`"en"`, default) and Brazilian Portuguese (`"br"`).}

\item{label}{Logical. If `TRUE` (default), variable labels are added to
the output data frame using `sjlabelled::set_label()`.}

\item{quiet}{Logical. If `FALSE` (default), a progress bar is displayed.}
}
\value{
A data frame containing the Ipeadata code, date, value,
  territorial unit name, and territorial code of the requested series.
}
\description{
Returns the data associated with requested Ipeadata series.
}
\note{
Ipeadata codes may be obtained using [available_series()].
}
\examples{
\donttest{
data <- ipeadata(
  code = "PRECOS12_IPCA12",
  language = "en"
)
}

}
\references{
This package uses the Ipeadata API:
  \url{https://www.ipeadata.gov.br/}
}
\seealso{
[available_series()], [available_territories()]
}
