% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipeadata_pkg.R
\name{available_countries}
\alias{available_countries}
\title{List with available countries}
\usage{
available_countries(language = c("en", "br"))
}
\arguments{
\item{language}{String specifying the selected language. Language options are
English (\code{"en"}, default) and Brazilian Portuguese (\code{"br"}).}
}
\value{
A data frame containing 3-letter country code and name of available countries.
}
\description{
Returns a list with available countries from Ipeadata API database.
}
