% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipeadata_pkg.R
\name{available_countries}
\alias{available_countries}
\title{List with countries' names}
\usage{
available_countries(language = c("en", "br"))
}
\arguments{
\item{language}{string specifying the selected language. Language options are
English (\code{"en"}, default) and Brazilian portuguese (\code{"br"}).}
}
\value{
A data frame containing the country code and the countries' names in Brazilian portuguese.
}
\description{
Returns a lista with all countries that Ipea API database has information about. (????)
}
\examples{
# Available countries (in English)
all_countries <- available_countries()

# Available countries (in Brazilian portuguese)
all_countriesBR <- available_countries(language = "br")

}
