% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipeadata_pkg.R
\name{available_series}
\alias{available_series}
\title{List with available series}
\usage{
available_series()
}
\description{
Returns the list with all available series from Ipeadata database
}
\examples{
# Available series
all_series <- available_series()

}
