% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipeadata_pkg.R
\name{available_series}
\alias{available_series}
\title{Available series}
\usage{
available_series(language = c("en", "br"))
}
\arguments{
\item{language}{A string specifying the language. Available options are
English (\code{"en"}, default) and Brazilian Portuguese (\code{"br"}).}
}
\value{
A data frame containing the Ipeadata code, name, theme, source,
  frequency, last update, and activity status of available series.
}
\description{
Returns a list of available series from the Ipeadata API.
}
\note{
The original language of the available series names is preserved.
}
