% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_subjects.R
\name{available_subjects}
\alias{available_subjects}
\title{Available Subjects}
\usage{
available_subjects(language = c("en", "br"), label = TRUE)
}
\arguments{
\item{language}{A character string specifying the language. Available
options are English (\code{"en"}, default) and Brazilian Portuguese (\code{"br"}).}

\item{label}{Logical. If \code{TRUE} (default), variable labels are added to
the output data frame using \code{sjlabelled::set_label()}.}
}
\value{
A data frame containing the subject code and subject name.
}
\description{
Returns a list of available subjects from the Ipeadata API.
}
