% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipeadata_pkg.R
\name{available_subjects}
\alias{available_subjects}
\title{List with available subjects}
\usage{
available_subjects(language = c("en", "br"))
}
\arguments{
\item{language}{String specifying the selected language. Language options are
English (\code{"en"}, default) and Brazilian portuguese (\code{"br"}).}
}
\value{
A data frame containing code and name of available subjects.
}
\description{
Returns a list with available subjects from Ipeadata API database.
}
\examples{
\donttest{
# Available subjects (in English)
all_subjects <- available_subjects()

# Available subjects (in Brazilian portuguese)
all_subjectsBR <- available_subjects(language = "br")

}

}
