% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipeadata_pkg.R
\name{available_subjects}
\alias{available_subjects}
\title{Available subjects}
\usage{
available_subjects(language = c("en", "br"))
}
\arguments{
\item{language}{A string specifying the language. Available options are
English (\code{"en"}, default) and Brazilian Portuguese (\code{"br"}).}
}
\value{
A data frame containing the code and name of available subjects.
}
\description{
Returns a list of available subjects from the Ipeadata API.
}
