% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_territories.R
\name{available_territories}
\alias{available_territories}
\title{Available Territorial Divisions}
\usage{
available_territories(language = c("en", "br"), label = TRUE)
}
\arguments{
\item{language}{A character string specifying the language. Available
options are English (\code{"en"}, default) and Brazilian Portuguese (\code{"br"}).}

\item{label}{Logical. If \code{TRUE} (default), variable labels are added to
the output data frame using \code{sjlabelled::set_label()}.}
}
\value{
A data frame containing the unit type, code, name, and area
(in km²) of Brazilian territorial divisions.
}
\description{
Returns a list of available Brazilian territorial divisions
from the Ipeadata API.
}
