% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipeadata_pkg.R
\name{available_territories}
\alias{available_territories}
\title{List with territories' names}
\usage{
available_territories(language = c("en", "br"))
}
\arguments{
\item{language}{string specifying the selected language. Language options are
English (\code{"en"}, default) and Brazilian portuguese (\code{"br"}).}
}
\value{
A data frame containing the territorial unit name, the territorial code, the territorial's name
and the area in km^2.
}
\description{
Returns a list containing information about territories' in Brazil.
}
\examples{
# Available territories (in English)
all_territories <- available_territories()

# Available territories (in Brazilian portuguese)
all_territoriesBR <- available_territories(language = "br")

}
