% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipeadata_pkg.R
\name{available_territories}
\alias{available_territories}
\title{List with available territorial divisions}
\usage{
available_territories(language = c("en", "br"))
}
\arguments{
\item{language}{String specifying the selected language. Language options are
English (\code{"en"}, default) and Brazilian portuguese (\code{"br"}).}
}
\value{
A data frame containing unit name, code, name and area (in km2)
 of Brazilian territorial divisions.
}
\description{
Returns a list with available Brazilian territorial
 divisions from Ipeadata API database.
}
\examples{
\donttest{
#' if (interactive()) {

# Available territories (in English)
all_territories <- available_territories()

# Available territories (in Brazilian portuguese)
all_territoriesBR <- available_territories(language = "br")

}
}

}
