% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipeadata_pkg.R
\name{available_territories}
\alias{available_territories}
\title{Available territorial divisions}
\usage{
available_territories(language = c("en", "br"))
}
\arguments{
\item{language}{A string specifying the language. Available options are
English (\code{"en"}, default) and Brazilian Portuguese (\code{"br"}).}
}
\value{
A data frame containing the unit type, code, name, and area (in km²)
  of Brazilian territorial divisions.
}
\description{
Returns a list of available Brazilian territorial divisions
from the Ipeadata API.
}
