% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipeadata.R
\name{ipeadata}
\alias{ipeadata}
\title{Data for Requested Series}
\usage{
ipeadata(code, language = c("en", "br"), label = TRUE, quiet = FALSE)
}
\arguments{
\item{code}{A character vector of Ipeadata series codes.}

\item{language}{A character string specifying the language. Available
options are English (\code{"en"}, default) and Brazilian Portuguese (\code{"br"}).}

\item{label}{Logical. If \code{TRUE} (default), variable labels are added to
the output data frame using \code{sjlabelled::set_label()}.}

\item{quiet}{Logical. If \code{FALSE} (default), a progress bar is displayed.}
}
\value{
A data frame containing the Ipeadata code, date, value,
territorial unit name, and territorial code of the requested series.
}
\description{
Returns the data associated with requested Ipeadata series.
}
\note{
Ipeadata codes may be obtained using \code{\link[=available_series]{available_series()}}.
}
\examples{
\donttest{data <- ipeadata(code = "PRECOS12_IPCA12")}
}
\references{
This package uses the Ipeadata API:
\url{https://www.ipeadata.gov.br/}
}
\seealso{
\code{\link[=available_series]{available_series()}}, \code{\link[=available_territories]{available_territories()}}
}
