% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipeadata_pkg.R
\name{ipeadata}
\alias{ipeadata}
\title{Returns a database about the chosen serie(s)}
\usage{
ipeadata(code, language = c("en", "br"), quiet = FALSE)
}
\arguments{
\item{code}{vector of the serie codes. This serie codes may be required by \code{available_series()}.}

\item{language}{string specifying the selected language. Language options are
English (\code{"en"}, default) and Brazilian portuguese (\code{"br"}).}

\item{quiet}{logical. If \code{quiet = FALSE} (default), a bar of progression is shown.}
}
\value{
A data frame containing the serie code, the date, the values, the territorial unit name
and the territorial/country code.
}
\description{
dfsdfsdfsdfdsfsd
}
\examples{
serieA <- ipeadata(c('abate_abpeav','AbInEe_VeLeTfC','VALOR366_FEDFUND366',
                              'ABCD', 'CONSUMOTOT', '13T_BANAGUA'))
serieB <- ipeadata(c('valor366_fedfund366'))
serieE <- ipeadata(c('abate_abpeav'))
serieF <- ipeadata(c('abate_abpeav'), language = c("br"))
serieG <- ipeadata(c('CONSUMOTOT'))
serieH <- ipeadata(c('CONSUMOTOT'), language = c("br"))
# serieC <- ipeadata(all_series$code)
serieD <- ipeadata(all_series$code[1:150])
serieD <- ipeadata(all_series$code[1:150], quiet = TRUE)

}
\references{

}
