% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipeadata_pkg.R
\name{ipeadata}
\alias{ipeadata}
\title{Returns a database about the requested series}
\usage{
ipeadata(code, language = c("en", "br"), quiet = FALSE)
}
\arguments{
\item{code}{A character vector with Ipeadata code.}

\item{language}{String specifying the selected language. Language options are
English (\code{"en"}, default) and Brazilian portuguese (\code{"br"}).}

\item{quiet}{Logical. If \code{FALSE} (default), a progress bar is shown.}
}
\value{
A data frame containing Ipeadata code, date, value, territorial unit name
and country or territorial code of requested series.
}
\description{
Returns a list with available database about the requested series.
}
\note{
The Ipeadata codes may be required by \code{available_series()}.
}
\references{
This R package uses the Ipeadata API.
For more information go to \url{http://www.ipeadata.gov.br/}.
}
\seealso{
\code{\link{available_series}}, \code{\link{available_territories}}
}
