% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipeadata_pkg.R
\name{ipeadata}
\alias{ipeadata}
\title{BUSCA VALORES}
\usage{
ipeadata(code, language = c("en", "br"), quiet = FALSE)
}
\arguments{
\item{code}{jhiuahuhsu}

\item{language}{string specifying the selected language. Language options are
English (\code{"en"}, default) and Brazilian portuguese (\code{"br"}).}

\item{quiet}{dijfisjdijfsid}
}
\value{
A data frame containing the series code, the series' name, a comment about the series, last uptade, , ,
the source, the source's name, the source's link, the frequency of the series, the unity, the multiplier factor,
the topic cod and the country/territorial code. dsfdsfdsfsrevcdxgdfgdrgdfdgdgfd
}
\description{
etfetrezetertatetdaertaertrrtaer
}
\examples{
serieA <- ipeadata(c('abate_abpeav','AbInEe_VeLeTfC','VALOR366_FEDFUND366',
                              'ABCD', 'CONSUMOTOT', '13T_BANAGUA'))
serieB <- ipeadata(c('valor366_fedfund366'))
serieE <- ipeadata(c('abate_abpeav'))
serieF <- ipeadata(c('abate_abpeav'), language = c("br"))
serieG <- ipeadata(c('CONSUMOTOT'))
serieH <- ipeadata(c('CONSUMOTOT'), language = c("br"))
# serieC <- ipeadata(all_series$code)
serieD <- ipeadata(all_series$code[1:150])
serieD <- ipeadata(all_series$code[1:150], TRUE)

}
