% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipeadata_pkg.R
\name{ipeadata}
\alias{ipeadata}
\title{Data for a requested series}
\usage{
ipeadata(code, language = c("en", "br"), quiet = FALSE)
}
\arguments{
\item{code}{A character vector of Ipeadata series codes.}

\item{language}{A string specifying the language. Available options are
English (\code{"en"}, default) and Brazilian Portuguese (\code{"br"}).}

\item{quiet}{Logical. If \code{FALSE} (default), a progress bar is displayed.}
}
\value{
A data frame containing the Ipeadata code, date, value,
  territorial unit name, and territorial code of the requested series.
}
\description{
Returns the data associated with the requested Ipeadata series.
}
\note{
The Ipeadata codes may be obtained using \code{available_series()}.
}
\examples{
\dontrun{
# Data for:
# "PRECOS12_IPCA12": Brazil’s official consumer price inflation index (IPCA)
data <- ipeadata(code = "PRECOS12_IPCA12", language = "en")
}

}
\references{
This package uses the Ipeadata API.
  For more information, see \url{https://www.ipeadata.gov.br/}.
}
\seealso{
\code{\link{available_series}},
  \code{\link{available_territories}}
}
