% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipeadatar-package.R
\docType{package}
\name{ipeadatar-package}
\alias{ipeadatar}
\alias{ipeadatar-package}
\title{ipeadatar: API Wrapper for Ipeadata}
\description{
Provides direct access to the macroeconomic, financial,
and regional database maintained by the Institute for
Applied Economic Research (Ipea) via the Ipeadata API.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ipea/ipeadatar}
  \item Report bugs at \url{https://github.com/ipea/ipeadatar/issues}
}

}
\author{
\strong{Maintainer}: Luiz Eduardo Gomes \email{luiz.gomes@ipea.gov.br} (\href{https://orcid.org/0000-0001-8968-5947}{ORCID})

Authors:
\itemize{
  \item Luiz Eduardo Gomes \email{luiz.gomes@ipea.gov.br} (\href{https://orcid.org/0000-0001-8968-5947}{ORCID})
}

Other contributors:
\itemize{
  \item Daniel Herszenhut \email{daniel.santos@ipea.gov.br} (\href{https://orcid.org/0000-0001-8066-1105}{ORCID}) [contributor]
  \item Instituto de Pesquisa Econômica Aplicada (Ipea) [copyright holder, funder]
}

}
\keyword{internal}
