% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipeadata_pkg.R
\name{metadata}
\alias{metadata}
\title{MetadadosXXXX}
\usage{
metadata(code, language = c("en", "br"), quiet = FALSE)
}
\arguments{
\item{code}{jhiuahuhsu}

\item{language}{string specifying the selected language. Language options are
English (\code{"en"}, default) and Brazilian portuguese (\code{"br"}).}

\item{quiet}{dijfisjdijfsid}
}
\value{
A data frame containing the series code, the series' name, a comment about the series, last uptade, , ,
the source, the source's name, the source's link, the frequency of the series, the unity, the multiplier factor,
the topic cod and the country/territorial code.
}
\description{
dsgfaggfedgaedgfewetdewategdg
}
\examples{
metadado.serieA <- metadata(c('abate_abpeav','AbInEe_VeLeTfC','VALOR366_FEDFUND366',
                              'ABCD', 'CONSUMOTOT', '13T_BANAGUA'))
metadado.serieB <- metadata(c('valor366_fedfund366'))
metadado.serieE <- metadata(c('abate_abpeav'))
metadado.serieF <- metadata(c('abate_abpeav'), language = c("br"))
# metadado.serieC <- metadata(all_series$code)
# metadado.serieD <- metadata(all_series$code[1:150])
metadado.serieD <- metadata(all_series$code[1:150], 'br' ,TRUE)

}
