% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipeadata_pkg.R
\name{metadata}
\alias{metadata}
\title{Returns a metadata about the chosen serie(s)}
\usage{
metadata(code, language = c("en", "br"), quiet = FALSE)
}
\arguments{
\item{code}{Vector of the serie codes. This serie codes may be required by \code{available_series()}.}

\item{language}{String specifying the selected language. Language options are
English (\code{"en"}, default) and Brazilian portuguese (\code{"br"}).}

\item{quiet}{Logical. If \code{quiet = FALSE} (default), a bar of progression is shown.}
}
\value{
A data frame containing the serie codes, the series' names, a comment about the series, last uptades,
the themes, the sources, the sources' names, the sources' url, the frequency of the series, the unity,
the multiplier factor, the series' status, the subject code and the country/territorial code.
}
\description{
Returns a list with information about the chosen serie(s).
}
\note{
The original language of the available series' names and the comments were preserved.
}
\examples{
metadado.serieA <- metadata(c('abate_abpeav','AbInEe_VeLeTfC','VALOR366_FEDFUND366',
                              'ABCD', 'CONSUMOTOT', '13T_BANAGUA'))
metadado.serieB <- metadata(c('valor366_fedfund366'))
metadado.serieE <- metadata(c('abate_abpeav'))
metadado.serieF <- metadata(c('abate_abpeav'), language = c("br"))
# metadado.serieC <- metadata(all_series$code)
# metadado.serieD <- metadata(all_series$code[1:150])
metadado.serieD <- metadata(all_series$code[1:150], 'br' ,TRUE)

metadado.serieF <- metadata(c('TMP30DJF'), language = c("br"))

}
