% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipeadata_pkg.R
\name{metadata}
\alias{metadata}
\title{Metadata for a requested series}
\usage{
metadata(code, language = c("en", "br"), quiet = FALSE)
}
\arguments{
\item{code}{A character vector of Ipeadata series codes.}

\item{language}{A string specifying the language. Available options are
English (\code{"en"}, default) and Brazilian Portuguese (\code{"br"}).}

\item{quiet}{Logical. If \code{FALSE} (default), a progress bar is displayed.}
}
\value{
A data frame containing the Ipeadata code, name, short comment,
  last update, theme name, source name and full name, source URL,
  frequency, unit, multiplier factor, status, subject code, and
  country code of the requested series.
}
\description{
Returns metadata information for the requested Ipeadata series.
}
\note{
The original language of the series names and comments is preserved.
  The Ipeadata codes may be obtained using \code{available_series()}.
}
\examples{
\dontrun{
# Metadata for:
# "PRECOS12_IPCA12": Brazil’s official consumer price inflation index (IPCA)
meta <- metadata(code = "PRECOS12_IPCA12")
}

}
\references{
This package uses the Ipeadata API.
  For more information, see \url{https://www.ipeadata.gov.br/}.
}
\seealso{
\code{\link{available_series}},
  \code{\link{available_subjects}},
  \code{\link{available_territories}}
}
