% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_series.R
\name{search_series}
\alias{search_series}
\title{Searched Series}
\usage{
search_series(terms = NULL, language = c("en", "br"), label = TRUE)
}
\arguments{
\item{terms}{A character vector of search terms.}

\item{language}{A character string specifying the language. Available
options are English (\code{"en"}, default) and Brazilian Portuguese (\code{"br"}).}

\item{label}{Logical. If \code{TRUE} (default), variable labels are added to
the output data frame using \code{sjlabelled::set_label()}.}
}
\value{
A data frame containing the Ipeadata code, name, theme, source,
frequency, last update, and activity status of the matched series.
}
\description{
Searches for Ipeadata series matching the provided terms.
}
\note{
The original language of the series names is preserved.
}
\examples{
\donttest{
search_series(terms = "rural")
}

}
