% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipeadata_pkg.R
\name{search_series}
\alias{search_series}
\title{Searched series}
\usage{
search_series(terms = NULL, language = c("en", "br"))
}
\arguments{
\item{terms}{A character vector of search terms.}

\item{language}{A string specifying the language. Available options are
English (\code{"en"}, default) and Brazilian Portuguese (\code{"br"}).}
}
\value{
A data frame containing the Ipeadata code, name, theme, source,
  frequency, last update, and activity status of the matched series.
}
\description{
Returns a list of series matching the specified search terms
  from the Ipeadata API.
}
\note{
The original language of the series names is preserved.
}
\examples{
\dontrun{
# Search for "rural"
search <- search_series(terms = "rural", language = "en")
}

}
